+++ packages/rx_bloc_cli/mason_templates/bricks/lib_dev_menu/__brick__/lib/lib_dev_menu/alice_instance.dart
@@ -1,8 +1,0 @@
-import 'package:alice/alice.dart';
-
-Alice alice = Alice(
-  showNotification: true,
-  showInspectorOnShake: false,
-  darkTheme: false,
-  maxCallsCount: 1000,
-);